Faster carrier detection
We now use a PCINT to detect a level change on the RX pin.  Thus we don't need to wait until the received received a complete and valid byte but we can reset the detection interval as soon as we see a change on the bus.

Also change the time measurement to use just Tc, the time it takes for
one octet on the wire.  The old code used 1.5*Tc.